requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.ini
#
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via -r requirements.ini
commonmark==0.9.1
    # via rich
deepl==1.11.0
    # via -r requirements.ini
idna==3.4
    # via requests
jinja2==3.1.2
    # via -r requirements.ini
markupsafe==2.1.1
    # via jinja2
pygments==2.13.0
    # via rich
python-dotenv==0.21.0
    # via -r requirements.ini
python-slugify==7.0.0
    # via -r requirements.ini
requests==2.28.1
    # via
    #   -r requirements.ini
    #   deepl
    #   tmdbv3api
rich==12.6.0
    # via -r requirements.ini
text-unidecode==1.3
    # via python-slugify
tmdbv3api==1.7.7
    # via -r requirements.ini
tvdb-v4-official==1.1.0
    # via -r requirements.ini
urllib3==1.26.13
    # via requests
